update: [정다운] 상품 삭제시 메인 이미지가 없어도 삭제할 수 있도록 변경 [SDY-138]

+++ src/main/java/com/dyes/backend/domain/product/service/admin/AdminProductServiceImpl.java
@@ -369,9 +369,10 @@ public class AdminProductServiceImpl implements AdminProductService {
             Optional<ProductMainImage> maybeProductMainImage = productMainImageRepository.findByProductWithProduct(deleteProduct);
             if (maybeProductMainImage.isEmpty()) {
                 log.info("ProductMainImage for product with ID '{}' is empty", deleteProduct.getId());
+            } else if(maybeProductMainImage.isPresent()){
+                ProductMainImage deleteProductMainImage = maybeProductMainImage.get();
+                productMainImageRepository.delete(deleteProductMainImage);
             }
-            ProductMainImage deleteProductMainImage = maybeProductMainImage.get();
-            productMainImageRepository.delete(deleteProductMainImage);
 
             List<ProductDetailImages> deleteProductDetailImagesList = productDetailImagesRepository.findByProductWithProduct(deleteProduct);
             for (ProductDetailImages productDetailImages : deleteProductDetailImagesList) {
@@ -381,9 +382,10 @@ public class AdminProductServiceImpl implements AdminProductService {
             Optional<ProductManagement> maybeProductManagement = productManagementRepository.findByProduct(deleteProduct);
             if(maybeProductManagement.isEmpty()) {
                 log.info("ProductManagement for product with ID '{}' is empty", deleteProduct.getId());
+            } else if(maybeProductMainImage.isPresent()){
+                ProductMainImage deleteProductMainImage = maybeProductMainImage.get();
+                productMainImageRepository.delete(deleteProductMainImage);
             }
-            ProductManagement productManagement = maybeProductManagement.get();
-            productManagementRepository.delete(productManagement);
 
             productRepository.delete(deleteProduct);
             log.info("Product deletion successful for product with ID: {}", productId);
